work around bug in shlex.split
not nice, but works.
Checking for the python version
and not converting if python >=2.7.3 might look strange to
the users where it works on one system but not on another ...

I chose cp1250 because I still know it from old dos times.
Would latin-1 be better ? I have no clue and thought this mess
would be over by now.