fix parsing of gpx files produced by garmin fenix watches
the fenix watch writes a wrong time in the metadata section of the gpx.
It is even in the future.
When using this for calculating the duration, you  get the time and
the wrong pace. I just ran 30km in 0.01 Minutes. That seems to be too fast :)

By using the time of the first trkseg as the starting time, everything
works as expected.