Do not log from Environment.__init__
Environment is created before running log_setup (and needs to be,
log_setup uses values from Environment). Outputting a log message
before logging is setup causes logging.basicConfig to run at default
values, and so the setup done in log_setup is ignored because logging
is already configured. The info message in Environment.__init__ was
introduced in 5a2292d.