Run tests in a separate container
The Github ubuntu-latest image is broken so we need to switch to a separate
image. The ubuntu-latest image has a bunch of packages installed and
settings changed by default, we need to follow a number of those while
making this change.

Also the database services are now available in separate hosts instead of
localhost (this is probably better).