Don't show heart rate graph for GPX files without HR data
It's wrong to assume that an activity has a heart rate graph to show just
because it has tracks.

There are 6 type of activities to consider:

1) With GPX file with HR data for (some of) the trackpoints
2) With GPX file with overall HR data but no trackpoints (indoor)
3) With GPX file without any HR data
4) With a GPX file as in point 3 with manually added HR data or calories
5) Without GPX file and with manually added HR data or calories
6) Without GPX file and without any HR data and calories

Only point 1 above have data for a heart rate graph out of the three with
GPX files that has tracks.

Currently it's possible to manually add HR data and calories to an activity
with a GPX file as in point 4 above on creation, but it's not possible to
edit the HR fields or calories after creation in an activity with a GPX file.